requirements.txt
torch
hydra-core
wandb
omegaconf